requirements.txt
flask
requests
python-dotenv
openai
playwright
pdfplumber
pandas
gunicorn